Fix errors on two pages of ES6 Tutorial
This commit incorporates the following fixes:

1. Rename webpack config entry from app to main

This fix renames the entry 'app' back to 'main' in webpack.config.js.
In webpack.config.js, the entries are defined as 'app' and
'ratefinder'. Due to this, the compiled js files will be app.bundle.js
and ratefinder.bundle.js,which  means the old main.bundle.js will never
be updated.

2. Add code for showing table to index.html

This commit adds a new point to part 1 of the Classes sub-tutorial,
which adds the missing <table> element to index.html. Without this,
the table will not be visible.